refactor(bot): extract exit execution into ExitExecutionMixin (phase 5)
First cut of splitting the 3.8k-line AuramaurBot god-class (the review's
"orchestrator too large" finding). Pure structural move, zero behavior change:
the poly / Kalshi / IBKR exit paths + helpers (_execute_*_exit,
_clear_exit_suppression, _prune_zero_onchain_poly_position, the lazy
_exit_gateway) move verbatim into auramaur/bot_exits.py as ExitExecutionMixin,
mixed into AuramaurBot. The methods still operate on the bot's self, so wiring,
components, and call sites are unchanged.

- Renamed the name-mangled self.__exit_gateway -> self._exit_gateway_obj so the
  lazy-gateway attribute (set in AuramaurBot.__init__) resolves across the mixin
  boundary (dunder mangling is per-class).
- Method-local imports moved with their methods; bot_exits adds only the
  module-level order types + logger it needs.

bot.py: 3819 -> 3462 lines. Parity proven by the existing exit + order-monitor
suites (all green, no test changes). Full suite: 837 passed.
